refactor: LocalStorage 폴더 이전
